// frontend/src/backend.d.ts
export enum IngredientSafety {
    safe = "safe",
    conditional = "conditional",
    harmful = "harmful",
    risky = "risky"
}
// ...
export enum Variant_fair_good_poor_unsafe_excellent {
    fair = "fair",
    good = "good",
    poor = "poor",
    unsafe = "unsafe",
    excellent = "excellent"
}